// packages/rum-core/src/rawRumEvent.types.ts
export const enum RumEventType {
  ACTION = 'action',
  ERROR = 'error',
  LONG_TASK = 'long_task',
  VIEW = 'view',
  RESOURCE = 'resource',
  VITAL = 'vital',
}

export const enum RumLongTaskEntryType {
  LONG_TASK = 'long-task',
  LONG_ANIMATION_FRAME = 'long-animation-frame',
}

export interface RawRumResourceEvent {
  date: TimeStamp
  type: RumEventType.RESOURCE
  resource: {
    type: ResourceType
    id: string
    duration?: ServerDuration
    url: string
    method?: string
    status_code?: number
    size?: number
    encoded_body_size?: number
    decoded_body_size?: number
    transfer_size?: number
    render_blocking_status?: string
    redirect?: ResourceEntryDetailsElement
    dns?: ResourceEntryDetailsElement
    connect?: ResourceEntryDetailsElement
    ssl?: ResourceEntryDetailsElement
    worker?: ResourceEntryDetailsElement
    first_byte?: ResourceEntryDetailsElement
    download?: ResourceEntryDetailsElement
    protocol?: string
    delivery_type?: DeliveryType
  }
  _dd: {
    trace_id?: string
    span_id?: string // not available for initial document tracing
    rule_psr?: number
    discarded: boolean
    page_states?: PageStateServerEntry[]
  }
}

export interface ResourceEntryDetailsElement {
  duration: ServerDuration
  start: ServerDuration
}
// ...
export type PageStateServerEntry = { state: PageState; start: ServerDuration }
// ...
export type DeliveryType = 'cache' | 'navigational-prefetch' | 'other'